Move metrics spi classes in the io.vertx.core.spi.metrics package
