+++ src/main/java/io/vertx/core/impl/VertxInternal.java
@@ -21,7 +21,7 @@ import io.netty.channel.EventLoopGroup;
 import io.vertx.core.Handler;
 import io.vertx.core.http.impl.HttpServerImpl;
 import io.vertx.core.json.JsonObject;
-import io.vertx.core.metrics.spi.VertxMetrics;
+import io.vertx.core.spi.metrics.VertxMetrics;
 import io.vertx.core.net.impl.NetServerImpl;
 import io.vertx.core.net.impl.ServerID;
 import io.vertx.core.spi.cluster.VertxSPI;
