+++ src/main/java/io/vertx/core/metrics/impl/DummyVertxMetrics.java
@@ -30,12 +30,12 @@ import io.vertx.core.http.HttpServerOptions;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.json.JsonObject;
-import io.vertx.core.metrics.spi.DatagramSocketMetrics;
-import io.vertx.core.metrics.spi.EventBusMetrics;
-import io.vertx.core.metrics.spi.HttpClientMetrics;
-import io.vertx.core.metrics.spi.HttpServerMetrics;
-import io.vertx.core.metrics.spi.NetMetrics;
-import io.vertx.core.metrics.spi.VertxMetrics;
+import io.vertx.core.spi.metrics.DatagramSocketMetrics;
+import io.vertx.core.spi.metrics.EventBusMetrics;
+import io.vertx.core.spi.metrics.HttpClientMetrics;
+import io.vertx.core.spi.metrics.HttpServerMetrics;
+import io.vertx.core.spi.metrics.NetMetrics;
+import io.vertx.core.spi.metrics.VertxMetrics;
 import io.vertx.core.net.NetClient;
 import io.vertx.core.net.NetClientOptions;
 import io.vertx.core.net.NetServer;
